+++ Source/Podio .NET/Services/CalendarService.cs
@@ -23,7 +23,7 @@ namespace PodioAPI.Services
         /// <param name="dateTo">The date to search to</param>
         /// <param name="priority">The minimum priority for the events to return Default value: 1</param>
         /// <returns></returns>
-        public async Task<IEnumerable<CalendarEvent>> GetAppCalendar(int appId, DateTime dateFrom, DateTime dateTo,
+        public async Task<IEnumerable<CalendarEvent>> GetAppCalendar(long appId, DateTime dateFrom, DateTime dateTo,
             int? priority = null)
         {
             string url = string.Format("/calendar/app/{0}/", appId);
@@ -65,7 +65,7 @@ namespace PodioAPI.Services
         /// <param name="userId"></param>
         /// <param name="token"></param>
         /// <returns></returns>
-        public async Task<StringResponse> GetAppCalendarAsiCal(int appId, int userId, string token)
+        public async Task<StringResponse> GetAppCalendarAsiCal(long appId, long userId, string token)
         {
             string url = string.Format("/calendar/app/{0}/ics/{1}/{2}/", appId, userId, token);
             return await _podio.Get<StringResponse>(url: url, returnAsString: true);
@@ -79,7 +79,7 @@ namespace PodioAPI.Services
         /// <param name="userId"></param>
         /// <param name="token"></param>
         /// <returns></returns>
-        public async Task<StringResponse> GetGlobalCalendarAsiCal(int userId, string token)
+        public async Task<StringResponse> GetGlobalCalendarAsiCal(long userId, string token)
         {
             string url = string.Format("/calendar/ics/{0}/{1}/", userId, token);
             return await _podio.Get<StringResponse>(url: url, returnAsString: true);
@@ -95,7 +95,7 @@ namespace PodioAPI.Services
         /// <param name="dateTo">The date to search to</param>
         /// <param name="priority">The minimum priority for the events to return Default value: 1</param>
         /// <returns></returns>
-        public async Task<IEnumerable<CalendarEvent>> GetSpaceCalendar(int spaceId, DateTime dateFrom, DateTime dateTo,
+        public async Task<IEnumerable<CalendarEvent>> GetSpaceCalendar(long spaceId, DateTime dateFrom, DateTime dateTo,
             int? priority = null)
         {
             string url = string.Format("/calendar/space/{0}/", spaceId);
@@ -117,7 +117,7 @@ namespace PodioAPI.Services
         /// <param name="userId"></param>
         /// <param name="token"></param>
         /// <returns></returns>
-        public async Task<StringResponse> GetSpaceCalendarAsiCal(int spaceId, int userId, string token)
+        public async Task<StringResponse> GetSpaceCalendarAsiCal(long spaceId, long userId, string token)
         {
             string url = string.Format("/calendar/space/{0}/ics/{1}/{2}/", spaceId, userId, token);
             var options = new Dictionary<string, bool>()
@@ -136,7 +136,7 @@ namespace PodioAPI.Services
         /// <param name="userId"></param>
         /// <param name="token"></param>
         /// <returns></returns>
-        public async Task<StringResponse> GetTaskCalendarAsiCal(int taskId)
+        public async Task<StringResponse> GetTaskCalendarAsiCal(long taskId)
         {
             string url = string.Format("/calendar/task/{0}/ics/", taskId);
             return await _podio.Get<StringResponse>(url: url, returnAsString: true);
@@ -188,7 +188,7 @@ namespace PodioAPI.Services
         /// <param name="limit">The maximum number of events to return in each group Default value: 5</param>
         /// <param name="priority">The minimum priority for the events to return Default value: 1</param>
         /// <returns></returns>
-        public async Task<CalendarSummary> GetCalendarSummaryForSpace(int spaceId, int limit = 5, int priority = 1)
+        public async Task<CalendarSummary> GetCalendarSummaryForSpace(long spaceId, int limit = 5, int priority = 1)
         {
             string url = string.Format("/calendar/space/{0}/summary", spaceId);
             var requestData = new Dictionary<string, string>
@@ -207,7 +207,7 @@ namespace PodioAPI.Services
         /// <param name="uid"></param>
         /// <param name="startDateTime"></param>
         /// <param name="endDateTime"></param>
-        public async Task<dynamic> UpdateCalendarEvent(int uid, DateTime startDateTime, DateTime endDateTime)
+        public async Task<dynamic> UpdateCalendarEvent(long uid, DateTime startDateTime, DateTime endDateTime)
         {
             string url = string.Format("/calendar/event/{0}", uid);
             dynamic requestData = new
