+++ Source/Podio .NET/Services/ConversationService.cs
@@ -20,7 +20,7 @@ namespace PodioAPI.Services
         /// </summary>
         /// <param name="conversationId"></param>
         /// <param name="participants">The list of people to grant access to. This is a list of contact identifiers</param>
-        public async Task<dynamic> AddParticipants(int conversationId, List<Ref> participants)
+        public async Task<dynamic> AddParticipants(long conversationId, List<Ref> participants)
         {
             string url = string.Format("/conversation/{0}/participant/v2/", conversationId);
             dynamic requestData = new
@@ -51,7 +51,7 @@ namespace PodioAPI.Services
         /// <param name="refId"></param>
         /// <param name="conversationCreateRequest"></param>
         /// <returns></returns>
-        public async Task<Conversation> CreateConversationOnObject(string refType, int refId,
+        public async Task<Conversation> CreateConversationOnObject(string refType, long refId,
             ConversationCreateRequest conversationCreateRequest)
         {
             string url = string.Format("/conversation/{0}/{1}/", refType, refId);
@@ -65,7 +65,7 @@ namespace PodioAPI.Services
         /// </summary>
         /// <param name="conversationId"></param>
         /// <returns></returns>
-        public async Task<Conversation> GetConversation(int conversationId)
+        public async Task<Conversation> GetConversation(long conversationId)
         {
             string url = string.Format("/conversation/{0}", conversationId);
             return await _podio.Get<Conversation>(url);
@@ -77,7 +77,7 @@ namespace PodioAPI.Services
         /// </summary>
         /// <param name="eventId"></param>
         /// <returns></returns>
-        public async Task<ConversationEvent> GetConversationEvent(int eventId)
+        public async Task<ConversationEvent> GetConversationEvent(long eventId)
         {
             string url = string.Format("/conversation/event/{0}", eventId);
             return await _podio.Get<ConversationEvent>(url);
@@ -91,7 +91,7 @@ namespace PodioAPI.Services
         /// <param name="limit">The maximum number of conversations/events to return. Default value: 10</param>
         /// <param name="offset">The offset into the list of conversations/events to return. Default value: 0</param>
         /// <returns></returns>
-        public async Task<List<ConversationEvent>> GetConversationEvents(int conversationId, int limit = 10, int offset = 0)
+        public async Task<List<ConversationEvent>> GetConversationEvents(long conversationId, int limit = 10, int offset = 0)
         {
             string url = string.Format("/conversation/{0}/event/", conversationId);
             var requestData = new Dictionary<string, string>()
@@ -127,7 +127,7 @@ namespace PodioAPI.Services
         /// <param name="refType"></param>
         /// <param name="refId"></param>
         /// <returns></returns>
-        public async Task<List<Conversation>> GetConversationsOnObject(string refType, int refId)
+        public async Task<List<Conversation>> GetConversationsOnObject(string refType, long refId)
         {
             string url = string.Format("/conversation/{0}/{1}/", refType, refId);
             return await _podio.Get<List<Conversation>>(url);
@@ -139,7 +139,7 @@ namespace PodioAPI.Services
         /// </summary>
         /// <param name="userId"></param>
         /// <returns></returns>
-        public async Task<Conversation> GetExistingDirectConversation(int userId)
+        public async Task<Conversation> GetExistingDirectConversation(long userId)
         {
             string url = string.Format("/conversation/direct/{0}", userId);
             return await _podio.Get<Conversation>(url);
@@ -182,7 +182,7 @@ namespace PodioAPI.Services
         ///     <para>Podio API Reference: https://developers.podio.com/doc/conversations/leave-conversation-35483748 </para>
         /// </summary>
         /// <param name="converstionId"></param>
-        public async Task<dynamic> LeaveConversation(int converstionId)
+        public async Task<dynamic> LeaveConversation(long converstionId)
         {
             string url = string.Format("/conversation/{0}/leave", converstionId);
             return await _podio.Post<dynamic>(url);
@@ -193,7 +193,7 @@ namespace PodioAPI.Services
         ///  <para>Podio API Reference: https://developers.podio.com/doc/conversations/mark-conversation-as-read-35441525 </para>
         /// </summary>
         /// <param name="conversationId"></param>
-        public async Task<dynamic> MarkConversationAsRead(int conversationId)
+        public async Task<dynamic> MarkConversationAsRead(long conversationId)
         {
             string url = string.Format("/conversation/{0}/read", conversationId);
             return await _podio.Post<dynamic>(url);
@@ -214,7 +214,7 @@ namespace PodioAPI.Services
         ///     <para>Podio API Reference: https://developers.podio.com/doc/conversations/mark-conversation-as-unread-35441542 </para>
         /// </summary>
         /// <param name="converstionId"></param>
-        public async Task<dynamic> MarkConversationAsUnread(int converstionId)
+        public async Task<dynamic> MarkConversationAsUnread(long converstionId)
         {
             string url = string.Format("/conversation/{0}/read", converstionId);
             return await _podio.Delete<dynamic>(url);
@@ -233,8 +233,8 @@ namespace PodioAPI.Services
         /// </param>
         /// <param name="embedUrl">The url to be attached</param>
         /// <returns></returns>
-        public async Task<ConversationEvent> ReplyToConversation(int converstionId, string text, List<int> fileIds = null,
-            int? embedId = null, string embedUrl = null)
+        public async Task<ConversationEvent> ReplyToConversation(long converstionId, string text, List<long> fileIds = null,
+            long? embedId = null, string embedUrl = null)
         {
             string url = string.Format("/conversation/{0}/reply/v2", converstionId);
             dynamic requestData = new
@@ -252,7 +252,7 @@ namespace PodioAPI.Services
         ///     <para>Podio API Reference: https://developers.podio.com/doc/conversations/star-conversation-35106944 </para>
         /// </summary>
         /// <param name="converstionId"></param>
-        public async Task<dynamic> StarConversation(int converstionId)
+        public async Task<dynamic> StarConversation(long converstionId)
         {
             string url = string.Format("/conversation/{0}/star", converstionId);
             return await _podio.Post<dynamic>(url);
@@ -263,7 +263,7 @@ namespace PodioAPI.Services
         ///     <para>Podio API Reference: https://developers.podio.com/doc/conversations/unstar-conversation-35106990 </para>
         /// </summary>
         /// <param name="converstionId"></param>
-        public async Task<dynamic> UnstarConversation(int converstionId)
+        public async Task<dynamic> UnstarConversation(long converstionId)
         {
             string url = string.Format("/conversation/{0}/star", converstionId);
             return await _podio.Delete<dynamic>(url);
